@@ -15,18 +15,18 @@
   },
   "main": "lib/index.js",
   "dependencies": {
-    "axios": "^1.7.2",
-    "firebase-admin": "^12.2.0",
-    "firebase-functions": "^5.0.1",
+    "axios": "^1.7.7",
+    "firebase-admin": "^12.4.0",
+    "firebase-functions": "^5.1.1",
     "node-telegram-bot-api": "^0.66.0",
-    "openai": "^4.52.7"
+    "openai": "^4.57.0"
   },
   "devDependencies": {
     "@types/node-telegram-bot-api": "^0.64.7", 
-    "eslint": "^9.7.0",
+    "eslint": "^9.9.1",
     "eslint-config-google": "^0.14.0",
     "firebase-functions-test": "^3.3.0",
-    "typescript": "^5.5.3"
+    "typescript": "^5.5.4"
   },
   "private": true
 }
