@@ -16,51 +16,51 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^18.0.4",
-    "@angular/common": "^18.0.4",
-    "@angular/compiler": "^18.0.4",
-    "@angular/core": "^18.0.4",
-    "@angular/forms": "^18.0.4",
-    "@angular/platform-browser": "^18.0.4",
-    "@angular/platform-browser-dynamic": "^18.0.4",
-    "@angular/platform-server": "^18.0.4",
-    "@angular/router": "^18.0.4",
-    "@angular/service-worker": "^18.0.4",
-    "@angular/ssr": "^18.0.5",
+    "@angular/animations": "^18.2.2",
+    "@angular/common": "^18.2.2",
+    "@angular/compiler": "^18.2.2",
+    "@angular/core": "^18.2.2",
+    "@angular/forms": "^18.2.2",
+    "@angular/platform-browser": "^18.2.2",
+    "@angular/platform-browser-dynamic": "^18.2.2",
+    "@angular/platform-server": "^18.2.2",
+    "@angular/router": "^18.2.2",
+    "@angular/service-worker": "^18.2.2",
+    "@angular/ssr": "^18.2.2",
     "express": "^4.19.2",
-    "ngx-count-animation": "^2.0.1",
+    "ngx-count-animation": "^2.0.2",
     "ngx-image-hero": "^2.0.2",
     "ngx-quicklink": "^0.4.6",
     "ngx-scroll-animations": "^2.0.1",
     "rxjs": "~7.8.1",
-    "tslib": "^2.6.3"
+    "tslib": "^2.7.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^18.0.5",
-    "@angular-eslint/builder": "18.0.1",
-    "@angular-eslint/eslint-plugin": "18.0.1",
-    "@angular-eslint/eslint-plugin-template": "18.0.1",
-    "@angular-eslint/schematics": "18.0.1",
-    "@angular-eslint/template-parser": "18.0.1",
-    "@angular/cli": "^18.0.5",
-    "@angular/compiler-cli": "^18.0.4",
-    "@angular/localize": "^18.0.4",
+    "@angular-devkit/build-angular": "^18.2.2",
+    "@angular-eslint/builder": "18.3.0",
+    "@angular-eslint/eslint-plugin": "18.3.0",
+    "@angular-eslint/eslint-plugin-template": "18.3.0",
+    "@angular-eslint/schematics": "18.3.0",
+    "@angular-eslint/template-parser": "18.3.0",
+    "@angular/cli": "^18.2.2",
+    "@angular/compiler-cli": "^18.2.2",
+    "@angular/localize": "^18.2.2",
     "@eslint/eslintrc": "^3.1.0",
-    "@eslint/js": "^9.9.0",
+    "@eslint/js": "^9.9.1",
     "@types/eslint__js": "^8.42.3",
     "@types/express": "^4.17.21",
     "@types/jasmine": "~5.1.4",
-    "@types/node": "^20.14.7",
+    "@types/node": "^22.5.2",
     "esbuild-visualizer": "^0.6.0",
-    "eslint": "^9.9.0",
-    "jasmine-core": "~5.1.2",
-    "karma": "~6.4.3",
+    "eslint": "^9.9.1",
+    "jasmine-core": "~5.2.0",
+    "karma": "~6.4.4",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "prettier": "^3.3.2",
-    "typescript": "~5.4.5",
-    "typescript-eslint": "^8.0.1"
+    "prettier": "^3.3.3",
+    "typescript": "~5.5.4",
+    "typescript-eslint": "^8.4.0"
   }
 }
