@@ -1,18 +1,19 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SharedTestingModule } from 'src/test/shared-testing.module';
-import { ProgressbarComponent } from './progressbar.component';
+import { ProgressBarComponent } from './progress-bar.component';
 
-describe('ProgressbarComponent', () => {
-  let component: ProgressbarComponent;
-  let fixture: ComponentFixture<ProgressbarComponent>;
+describe('ProgressBarComponent', () => {
+  let component: ProgressBarComponent;
+  let fixture: ComponentFixture<ProgressBarComponent>;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [ProgressbarComponent, SharedTestingModule],
+      imports: [ProgressBarComponent, SharedTestingModule],
     }).compileComponents();
 
-    fixture = TestBed.createComponent(ProgressbarComponent);
+    fixture = TestBed.createComponent(ProgressBarComponent);
+    fixture.componentRef.setInput('progress', 50);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
