@@ -1,8 +1,6 @@
 import {
-    ChangeDetectionStrategy,
-    Component,
-    ViewChild,
-    ViewContainerRef
+  ChangeDetectionStrategy,
+  Component
 } from '@angular/core';
 import { NgxScrollAnimationsDirective } from 'ngx-scroll-animations';
 import { SERVICES } from 'src/app/configs/services';
@@ -19,10 +17,6 @@ import { ServiceCard } from './model/service-card-interface';
   styleUrl: './about-me-services.component.scss',
 })
 export class AboutMeServicesComponent extends ExtendedComponent {
-  /** Reference to the container where service items will be created dynamically. */
-  @ViewChild('serviceItemsRef', { read: ViewContainerRef, static: true })
-  serviceItemsRef!: ViewContainerRef;
-
   /** Array of service card data to be displayed. */
-  public items: ServiceCard[] = SERVICES;
+  public readonly items: ServiceCard[] = SERVICES;
 }
