@@ -4,18 +4,16 @@ import {
   Component,
   OnInit,
   TemplateRef,
-  ViewChild,
   ViewContainerRef,
-  inject,
   signal,
+  viewChild
 } from '@angular/core';
 import { NgxImageHeroDirective } from 'ngx-image-hero';
 import { filter, fromEvent, timer } from 'rxjs';
 import { modalAnimation } from 'src/app/animations/modal-animations';
 import { ImageLoaderDirective } from 'src/app/directives/image-loader.directive';
 import { SafePipe } from 'src/app/pipes/safe.pipe';
-import { ModalManagerService } from 'src/app/services/modal-manager/modal-manager.service';
-import { ExtendedComponent } from 'src/app/utils/extended-component';
+import { Modal } from 'src/app/utils/modal/modal.component';
 import {
   BadgeTemplateI,
   ProjectDetails,
@@ -38,29 +36,33 @@ import {
   styleUrl: './project-details.component.scss',
   animations: [modalAnimation],
 })
-export class ProjectDetailsComponent
-  extends ExtendedComponent
-  implements OnInit
-{
+export class ProjectDetailsComponent extends Modal<ProjectDetails> implements OnInit {
   /** Reference to the container for displaying website URLs. */
-  @ViewChild('websiteContainer', { read: ViewContainerRef, static: true })
-  websiteContainer!: ViewContainerRef;
+  private websiteContainer = viewChild.required('websiteContainer', {
+    read: ViewContainerRef,
+  });
   /** Reference to the container for displaying technology badges. */
-  @ViewChild('technologyContainer', { read: ViewContainerRef, static: true })
-  technologyContainer!: ViewContainerRef;
+  private technologyContainer = viewChild.required('technologyContainer', {
+    read: ViewContainerRef,
+  });
   /** Reference to the container for displaying video player. */
-  @ViewChild('videoContainer', { read: ViewContainerRef, static: true })
-  videoContainer!: ViewContainerRef;
+  private videoContainer = viewChild.required('videoContainer', {
+    read: ViewContainerRef,
+  });
 
   /** Reference to the badge template. */
-  @ViewChild('badgeTemplate', { read: TemplateRef, static: true })
-  badgeTemplate!: TemplateRef<{ title: string; items: BadgeTemplateI[] }>;
+  private badgeTemplate = viewChild.required('badgeTemplate', {
+    read: TemplateRef<{ title: string; items: BadgeTemplateI[] }>,
+  });
   /** Reference to the URL list template. */
-  @ViewChild('urlListTemplate', { read: TemplateRef, static: true })
-  urlListTemplate!: TemplateRef<{ title: string; items: UrlListTemplateI[] }>;
+  private urlListTemplate = viewChild.required('urlListTemplate', {
+    read: TemplateRef<{ title: string; items: UrlListTemplateI[] }>,
+  });
   /** Reference to the YouTube player template. */
-  @ViewChild('youtubePlayer', { read: TemplateRef, static: true })
-  youtubePlayer!: TemplateRef<{ url: string }>;
+  private youtubePlayer = viewChild.required('youtubePlayer', {
+    read: TemplateRef<{ url: string }>,
+  });
+
 
   /** Duration of modal animation for fade in. */
   public readonly modalAnimationDurationIn = signal(500);
@@ -74,8 +76,6 @@ export class ProjectDetailsComponent
   public copiedInstallCode = signal(false);
   /** Flag indicating if "Copied!" message should be shown. */
   public showCopiedMsg = signal(false);
-  /** Modal manager service for managing modals. */
-  private modalManager = inject(ModalManagerService);
 
   public openHero = signal(false);
 
@@ -92,10 +92,10 @@ export class ProjectDetailsComponent
    * with the corresponding data from the `data` object.
    */
   private createDetailInfos() {
-    if (this.data.website) {
-      this.websiteContainer.createEmbeddedView(this.urlListTemplate, {
+    if (this.data().website) {
+      this.websiteContainer().createEmbeddedView(this.urlListTemplate(), {
         title: $localize`Website`,
-        items: this.data.website.map((el) => {
+        items: this.data().website!.map((el) => {
           return {
             ...el,
             name: el.url,
@@ -103,10 +103,10 @@ export class ProjectDetailsComponent
         }),
       });
     }
-    if (this.data.store) {
-      this.websiteContainer.createEmbeddedView(this.urlListTemplate, {
+    if (this.data().store) {
+      this.websiteContainer().createEmbeddedView(this.urlListTemplate(), {
         title: $localize`Mobile-Store`,
-        items: this.data.store.map((el) => {
+        items: this.data().store!.map((el) => {
           return {
             ...el,
             name: el.title,
@@ -115,14 +115,14 @@ export class ProjectDetailsComponent
       });
     }
 
-    if (this.data.video) {
-      this.videoContainer.createEmbeddedView(this.youtubePlayer, {
-        url: this.data.video,
+    if (this.data().video) {
+      this.videoContainer().createEmbeddedView(this.youtubePlayer(), {
+        url: this.data().video,
       });
     }
 
-    if (this.data.technology) {
-      Object.keys(this.data.technology).map((key) => {
+    if (this.data().technology) {
+      Object.keys(this.data().technology).map((key) => {
         const title =
           key === 'frontend'
             ? $localize`Frontend`
@@ -135,9 +135,9 @@ export class ProjectDetailsComponent
                   : null;
         if (!title) throw new Error(`Title for key '${key}' does not exists!`);
 
-        this.technologyContainer.createEmbeddedView(this.badgeTemplate, {
+        this.technologyContainer().createEmbeddedView(this.badgeTemplate(), {
           title,
-          items: this.data.technology[key as 'frontend'],
+          items: this.data().technology[key as 'frontend'],
         });
       });
     }
@@ -164,7 +164,7 @@ export class ProjectDetailsComponent
     timer(this.modalAnimationDurationOut())
       .pipe(this.destroyPipe())
       .subscribe(() => {
-        this.modalManager.closeModal();
+        this.close();
       });
   }
 
@@ -186,10 +186,5 @@ export class ProjectDetailsComponent
       .subscribe(() => {
         this.showCopiedMsg.set(false);
       });
-  }
-
-  /** Gets the project details data. */
-  public get data(): ProjectDetails {
-    return this.modalManager.modalData;
   }
 }
