@@ -1,42 +1,33 @@
+import { Renderer2, ViewContainerRef } from '@angular/core';
 import { TestBed } from '@angular/core/testing';
-
+import { renderer2Mock } from 'src/test/mocks/renderer2.mock';
+import { viewContainerRefMock } from 'src/test/mocks/view-container-ref.mock';
 import { SharedTestingModule } from 'src/test/shared-testing.module';
-import { ModalManagerService } from './modal-manager.service';
+import { ModalManager } from './modal-manager.service';
 
 describe('ModalManagerService', () => {
-  let service: ModalManagerService;
+  let service: ModalManager;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [SharedTestingModule],
-      teardown: {destroyAfterEach: false} 
+      providers: [
+        {
+          provide: Renderer2,
+          useValue: renderer2Mock,
+        },
+        {
+          provide: ViewContainerRef,
+          useValue: viewContainerRefMock,
+        },
+        ModalManager,
+      ],
+      teardown: { destroyAfterEach: false },
     });
-    service = TestBed.inject(ModalManagerService);
+    service = TestBed.inject(ModalManager);
   });
 
   it('should be created', () => {
     expect(service).toBeTruthy();
-  });
-
-  it('should open modal', () => {
-    const mockComponent = { name: 'MockComponent' };
-    const mockData = { data: { id: 123 } };
-
-    service.openModal(mockComponent, mockData);
-
-    expect(service.modalData).toEqual(mockData.data);
-    service.modal$.subscribe((modal) => {
-      expect(modal.type).toBe('add');
-      expect(modal.component).toBe(mockComponent);
-    });
-  });
-
-  it('should close modal', () => {
-    service.closeModal();
-
-    expect(service.modalData).toBeUndefined();
-    service.modal$.subscribe((modal) => {
-      expect(modal.type).toBe('remove');
-    });
   });
 });
