@@ -2,7 +2,7 @@ import { DOCUMENT } from '@angular/common';
 import { Injectable, inject, signal } from '@angular/core';
 import { Subject } from 'rxjs';
 import { getTheme } from '../../layout/header/components/theme-switch/utils/theme-switch';
-import { IS_BROWSER } from '../../utils/is-browser.provider';
+import { IS_BROWSER } from '../../utils/providers/is-browser.provider';
 
 @Injectable({
   providedIn: 'root',
