@@ -12,7 +12,7 @@ import { AnalyticsService } from '../services/analytics/analytics.service';
 import { MetaManagerService } from '../services/meta-manager/meta-manager.service';
 import { PageMetaData } from '../services/meta-manager/page-meta-data.interface';
 import { ScreenService } from '../services/screen/screen.service';
-import { IS_BROWSER } from './is-browser.provider';
+import { IS_BROWSER } from './providers/is-browser.provider';
 
 @Directive() // Dummy decorator
 /**
